requirements.txt
matplotlib==3.10.0
numpy==1.23.5
opencv_python==4.10.0.84
Pillow==11.1.0

